Add missing any/all option to stream rules in routing/intake page (#20678) (#20712)
* add missing any/all option to stream rules in Routing/intake page

* changelog file

* update MatchingTypeSwitcher component

* fix eslint issues
